Adds ProgressFileWriter class for file operations within ProgressWriter
